Change the way files are saved and pushed
saveFile was previously saving the file and pushing it to the remote repository.
Since git is not able to work in parallel (a lock file prevents it) it prevents us from saving several files at once.

saveFile now only saves the contents of the file.
A new endpoint, :owner/:repo/flush, was added to create the commits and push them to the remote repository.