Add matches{App, Project}Hostname to environment
These functions are responsible to match a specific hostname with the
app or project hostname. They return true when the hostname given is
the same of a subdomain of the app or project hostname.

In development mode (non-production) the project hostname will match
even when the xip ip is a different one. This is useful to access the
preview from a remote computer.