Make session.githubUser a promise
So that we first encounter a new session we don't block the request
waiting for Github to respond, and instead only when we actually need
details from the githubUser object.

In practice this makes POSTing the session to the project server a lot
faster, along with when we restart the login server.