Make the npm run *-debug commands work consistently
Turns out pkill was matching the `sudo` command that was running
pkill, as well as the server itself. This would cause the command to
exit with a non-zero status most of the time, occasionally the process
wouldn't get matched allowing it to work!

This change specifies that the process must be owned by the `montage`
user, which will only match the server.