Aggregate object functions in MontageObject
Replace findObjects with MontageObject.findAll.
Start migrating the code to a OO architecture.
I now have a better understanding of what we need in the LiveEdit so
I'm creating a "class" with the functions used to manipulate objects.
This is still work in progress and the code will be transitioned on a
need basis.
MontageElement, MontageComponent and MontageScope will have to follow.