Remove setObjectTemplate
We are now manually changing the Template on the Preview side.
The tool only changes the template when a library item is added but we
need to change in many more situations like a property change or new
binding. Because of this template changes done by the tool will be
wrong because they're missing all the other changes.