Invalidate all repetitions with zero iterations
Since we're not able to find which iteration templates need to be
invalidated when we can't find live elements to update
 - addTemplateFragment - we invalidate all repetitions with zero
iterations that are found inside the component we're adding the
TemplateFragment.
This will invalidate more repetitions than needed but at the moment
it's the best heuristic we have, we can come up with something better
in the future.