Revert "Stop committing and pushing on save"
I'm putting the actual commiting back where it belongs on the
environment and the backend. While filament's projectDocument
should coordinate the act of saving and offer some insight
into the state of things, it shouldn't bother being involved
in the actual defintion of what it means to "save" in a
given environment.

This reverts commit 235159483176cd1c183628b788ffad99e4da6f0b.

Conflicts:
	inject/adaptor/client/core/environment-bridge.js