Push the current branch
This was restricted to pushing a specific branch. After today we need
to relocate all git operations into repositoryService and make the
projectWorkspace rely on that API instead of the lower-level git
API directly.

This commit simply pushes the current branch you're on; if it happens
to be set as a tracking branch it works fine.