Load a template when making changes to it
We were previously getting an instance of a component and reaching for
its template instance.
The problem with this approach was that for components that didn't have
instances in the application template changes would not be applied and
lost forever.
New instances of these components will still get the old version of the
template.