Apply setObjectProperties changes to templates too
Given the huge volume of operations (60ops/s) of this kind fueled by the
 flow-editor we were avoiding changing the template on each change for
performance reasons.
This created the problem of us losing these changes if they were applied
to an object (a flow in this case) that wasn't instantiated in the app.
The strategy adopted to make this change is to wait until no more changes
of this type to the same object were performed in a specific amount of
time. This ensures that we don't save the template while we're
processing a stream of this type of operation.